requirements.txt
Flask
numpy
pandas
joblib
tensorflow-cpu
scikit-learn
gunicorn